+++ products/src/main/java/com/inventory/products/model/Product.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,7 +16,7 @@ public class Product {
     private String id;
     private String name;
     private Category category;
-    private float unitPrice;
+    private BigDecimal unitPrice;
     private LocalDate expirationDate;
     private int inStock;
     private LocalDate createdAt;
